# sms_api/api/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: sms-api
  version: 0.2.20-dev
paths:
...
  /configs/upload/simulation:
    post:
      tags:
      - Configurations - vEcoli
      summary: Upload Simulation Config
      description: Upload simulation config JSON
      operationId: upload-simulation-config
      parameters:
      - name: config_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Config Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationConfiguration'
              default:
                experiment_id: <PLACEHOLDER>
                suffix_time: true
                parca_options:
                  cpus: 2
                  outdir: /home/FCAM/svc_vivarium/workspace/outputs
                  operons: true
                  ribosome_fitting: true
                  rnapoly_fitting: true
                  remove_rrna_operons: false
                  remove_rrff: false
                  stable_rrna: false
                  new_genes: 'off'
                  debug_parca: false
                  save_intermediates: false
                  intermediates_directory: ''
                  variable_elongation_transcription: true
                  variable_elongation_translation: false
                generations: 1
                n_init_sims: 1
                max_duration: 10800.0
                initial_global_time: 0.0
                time_step: 1.0
                single_daughters: true
                emitter: parquet
                emitter_arg:
                  out_dir: /home/FCAM/svc_vivarium/workspace/outputs
                variants: {}
                analysis_options:
                  single:
                    mass_fraction_summary: {}
                    ptools_rxns:
                      n_tp: 8
                    ptools_rna:
                      n_tp: 8
                    ptools_proteins:
                      n_tp: 8
                agent_id: '0'
                parallel: false
                divide: true
                d_period: true
                division_threshold: true
                division_variable:
                - divide
                chromosome_path:
                - unique
                - full_chromosome
                spatial_environment: false
                fixed_media: minimal
                condition: basal
                save: false
                save_times: []
                add_processes: []
                exclude_processes: []
                profile: false
                processes:
                - post-division-mass-listener
                - bulk-timeline
                - media_update
                - exchange_data
                - ecoli-tf-unbinding
                - ecoli-equilibrium
                - ecoli-two-component-system
                - ecoli-rna-maturation
                - ecoli-tf-binding
                - ecoli-transcript-initiation
                - ecoli-polypeptide-initiation
                - ecoli-chromosome-replication
                - ecoli-protein-degradation
                - ecoli-rna-degradation
                - ecoli-complexation
                - ecoli-transcript-elongation
                - ecoli-polypeptide-elongation
                - ecoli-chromosome-structure
                - ecoli-metabolism
                - ecoli-mass-listener
                - RNA_counts_listener
                - rna_synth_prob_listener
                - monomer_counts_listener
                - dna_supercoiling_listener
                - replication_data_listener
                - rnap_data_listener
                - unique_molecule_counts
                - ribosome_data_listener
                - global_clock
                process_configs:
                  global_clock: {}
                  replication_data_listener:
                    time_step: 1
                topology:
                  bulk-timeline:
                    bulk:
                    - bulk
                    global:
                    - timeline
                    media_id:
                    - environment
                    - media_id
                  global_clock:
                    global_time:
                    - global_time
                    next_update_time:
                    - next_update_time
                engine_process_reports:
                - - listeners
                emit_paths: []
                progress_bar: true
                emit_topology: false
                emit_processes: false
                emit_config: false
                emit_unique: false
                log_updates: false
                raw_output: true
                description: ''
                seed: 0
                mar_regulon: false
                amp_lysis: false
                initial_state_file: ''
                skip_baseline: false
                daughter_outdir: /home/FCAM/svc_vivarium/workspace/outputs
                lineage_seed: 0
                fail_at_max_duration: false
                inherit_from: []
                spatial_environment_config: {}
                swap_processes: {}
                flow:
                  post-division-mass-listener: []
                  media_update:
                  - - post-division-mass-listener
                  exchange_data:
                  - - media_update
                  ecoli-tf-unbinding:
                  - - media_update
                  ecoli-equilibrium:
                  - - ecoli-tf-unbinding
                  ecoli-two-component-system:
                  - - ecoli-tf-unbinding
                  ecoli-rna-maturation:
                  - - ecoli-tf-unbinding
                  ecoli-tf-binding:
                  - - ecoli-equilibrium
                  ecoli-transcript-initiation:
                  - - ecoli-tf-binding
                  ecoli-polypeptide-initiation:
                  - - ecoli-tf-binding
                  ecoli-chromosome-replication:
                  - - ecoli-tf-binding
                  ecoli-protein-degradation:
                  - - ecoli-tf-binding
                  ecoli-rna-degradation:
                  - - ecoli-tf-binding
                  ecoli-complexation:
                  - - ecoli-tf-binding
                  ecoli-transcript-elongation:
                  - - ecoli-complexation
                  ecoli-polypeptide-elongation:
                  - - ecoli-complexation
                  ecoli-chromosome-structure:
                  - - ecoli-polypeptide-elongation
                  ecoli-metabolism:
                  - - ecoli-chromosome-structure
                  ecoli-mass-listener:
                  - - ecoli-metabolism
                  RNA_counts_listener:
                  - - ecoli-metabolism
                  rna_synth_prob_listener:
                  - - ecoli-metabolism
                  monomer_counts_listener:
                  - - ecoli-metabolism
                  dna_supercoiling_listener:
                  - - ecoli-metabolism
                  replication_data_listener:
                  - - ecoli-metabolism
                  rnap_data_listener:
                  - - ecoli-metabolism
                  unique_molecule_counts:
                  - - ecoli-metabolism
                  ribosome_data_listener:
                  - - ecoli-metabolism
                initial_state_overrides: []
                initial_state: {}
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadedSimulationConfig'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /configs/upload/analysis:
    post:
      tags:
      - Configurations - vEcoli
      summary: Upload Analysis Config
      description: Upload analysis config JSON
      operationId: upload-analysis-config
      parameters:
      - name: config_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Config Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationConfiguration'
              default:
                experiment_id: <PLACEHOLDER>
                suffix_time: true
                parca_options:
                  cpus: 2
                  outdir: /home/FCAM/svc_vivarium/workspace/outputs
                  operons: true
                  ribosome_fitting: true
                  rnapoly_fitting: true
                  remove_rrna_operons: false
                  remove_rrff: false
                  stable_rrna: false
                  new_genes: 'off'
                  debug_parca: false
                  save_intermediates: false
                  intermediates_directory: ''
                  variable_elongation_transcription: true
                  variable_elongation_translation: false
                generations: 1
                n_init_sims: 1
                max_duration: 10800.0
                initial_global_time: 0.0
                time_step: 1.0
                single_daughters: true
                emitter: parquet
                emitter_arg:
                  out_dir: /home/FCAM/svc_vivarium/workspace/outputs
                variants: {}
                analysis_options:
                  single:
                    mass_fraction_summary: {}
                    ptools_rxns:
                      n_tp: 8
                    ptools_rna:
                      n_tp: 8
                    ptools_proteins:
                      n_tp: 8
                agent_id: '0'
                parallel: false
                divide: true
                d_period: true
                division_threshold: true
                division_variable:
                - divide
                chromosome_path:
                - unique
                - full_chromosome
                spatial_environment: false
                fixed_media: minimal
                condition: basal
                save: false
                save_times: []
                add_processes: []
                exclude_processes: []
                profile: false
                processes:
                - post-division-mass-listener
                - bulk-timeline
                - media_update
                - exchange_data
                - ecoli-tf-unbinding
                - ecoli-equilibrium
                - ecoli-two-component-system
                - ecoli-rna-maturation
                - ecoli-tf-binding
                - ecoli-transcript-initiation
                - ecoli-polypeptide-initiation
                - ecoli-chromosome-replication
                - ecoli-protein-degradation
                - ecoli-rna-degradation
                - ecoli-complexation
                - ecoli-transcript-elongation
                - ecoli-polypeptide-elongation
                - ecoli-chromosome-structure
                - ecoli-metabolism
                - ecoli-mass-listener
                - RNA_counts_listener
                - rna_synth_prob_listener
                - monomer_counts_listener
                - dna_supercoiling_listener
                - replication_data_listener
                - rnap_data_listener
                - unique_molecule_counts
                - ribosome_data_listener
                - global_clock
                process_configs:
                  global_clock: {}
                  replication_data_listener:
                    time_step: 1
                topology:
                  bulk-timeline:
                    bulk:
                    - bulk
                    global:
                    - timeline
                    media_id:
                    - environment
                    - media_id
                  global_clock:
                    global_time:
                    - global_time
                    next_update_time:
                    - next_update_time
                engine_process_reports:
                - - listeners
                emit_paths: []
                progress_bar: true
                emit_topology: false
                emit_processes: false
                emit_config: false
                emit_unique: false
                log_updates: false
                raw_output: true
                description: ''
                seed: 0
                mar_regulon: false
                amp_lysis: false
                initial_state_file: ''
                skip_baseline: false
                daughter_outdir: /home/FCAM/svc_vivarium/workspace/outputs
                lineage_seed: 0
                fail_at_max_duration: false
                inherit_from: []
                spatial_environment_config: {}
                swap_processes: {}
                flow:
                  post-division-mass-listener: []
                  media_update:
                  - - post-division-mass-listener
                  exchange_data:
                  - - media_update
                  ecoli-tf-unbinding:
                  - - media_update
                  ecoli-equilibrium:
                  - - ecoli-tf-unbinding
                  ecoli-two-component-system:
                  - - ecoli-tf-unbinding
                  ecoli-rna-maturation:
                  - - ecoli-tf-unbinding
                  ecoli-tf-binding:
                  - - ecoli-equilibrium
                  ecoli-transcript-initiation:
                  - - ecoli-tf-binding
                  ecoli-polypeptide-initiation:
                  - - ecoli-tf-binding
                  ecoli-chromosome-replication:
                  - - ecoli-tf-binding
                  ecoli-protein-degradation:
                  - - ecoli-tf-binding
                  ecoli-rna-degradation:
                  - - ecoli-tf-binding
                  ecoli-complexation:
                  - - ecoli-tf-binding
                  ecoli-transcript-elongation:
                  - - ecoli-complexation
                  ecoli-polypeptide-elongation:
                  - - ecoli-complexation
                  ecoli-chromosome-structure:
                  - - ecoli-polypeptide-elongation
                  ecoli-metabolism:
                  - - ecoli-chromosome-structure
                  ecoli-mass-listener:
                  - - ecoli-metabolism
                  RNA_counts_listener:
                  - - ecoli-metabolism
                  rna_synth_prob_listener:
                  - - ecoli-metabolism
                  monomer_counts_listener:
                  - - ecoli-metabolism
                  dna_supercoiling_listener:
                  - - ecoli-metabolism
                  replication_data_listener:
                  - - ecoli-metabolism
                  rnap_data_listener:
                  - - ecoli-metabolism
                  unique_molecule_counts:
                  - - ecoli-metabolism
                  ribosome_data_listener:
                  - - ecoli-metabolism
                initial_state_overrides: []
                initial_state: {}
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadedAnalysisConfig'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /configs/get:
    get:
      tags:
      - Configurations - vEcoli
      summary: Get Simulation Config
      operationId: get-simulation-config
      parameters:
      - name: config_id
        in: query
        required: true
        schema:
          type: string
          title: Config Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationConfiguration'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /configs/remove:
    delete:
      tags:
      - Configurations - vEcoli
      summary: Delete Simulation Config
      operationId: delete-simulation-config
      parameters:
      - name: config_id
        in: query
        required: true
        schema:
          type: string
          title: Config Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Delete-Simulation-Config
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SimulationConfiguration:
      properties:
        experiment_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Experiment Id
        sim_data_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Sim Data Path
        suffix_time:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Suffix Time
        parca_options:
          additionalProperties: true
          type: object
          title: Parca Options
        generations:
          anyOf:
          - type: integer
          - type: 'null'
          title: Generations
        n_init_sims:
          anyOf:
          - type: integer
          - type: 'null'
          title: N Init Sims
        max_duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Max Duration
        initial_global_time:
          anyOf:
          - type: number
          - type: 'null'
          title: Initial Global Time
        time_step:
          anyOf:
          - type: number
          - type: 'null'
          title: Time Step
        single_daughters:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Single Daughters
        emitter:
          anyOf:
          - type: string
          - type: 'null'
          title: Emitter
        emitter_arg:
          additionalProperties: true
          type: object
          title: Emitter Arg
        variants:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Variants
        analysis_options:
          additionalProperties: true
          type: object
          title: Analysis Options
        gcloud:
          anyOf:
          - type: string
          - type: 'null'
          title: Gcloud
        agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Id
        parallel:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Parallel
        divide:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Divide
        d_period:
          anyOf:
          - type: boolean
          - type: 'null'
          title: D Period
        division_threshold:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Division Threshold
        division_variable:
          items:
            type: string
          type: array
          title: Division Variable
        chromosome_path:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Chromosome Path
        spatial_environment:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Spatial Environment
        fixed_media:
          anyOf:
          - type: string
          - type: 'null'
          title: Fixed Media
        condition:
          anyOf:
          - type: string
          - type: 'null'
          title: Condition
        save:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Save
        save_times:
          items:
            type: string
          type: array
          title: Save Times
        add_processes:
          items:
            type: string
          type: array
          title: Add Processes
        exclude_processes:
          items:
            type: string
          type: array
          title: Exclude Processes
        profile:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Profile
        processes:
          items:
            type: string
          type: array
          title: Processes
        process_configs:
          additionalProperties: true
          type: object
          title: Process Configs
        topology:
          additionalProperties: true
          type: object
          title: Topology
        engine_process_reports:
          items:
            items:
              type: string
            type: array
          type: array
          title: Engine Process Reports
        emit_paths:
          items:
            type: string
          type: array
          title: Emit Paths
        progress_bar:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Progress Bar
        emit_topology:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Emit Topology
        emit_processes:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Emit Processes
        emit_config:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Emit Config
        emit_unique:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Emit Unique
        log_updates:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Log Updates
        raw_output:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Raw Output
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
        mar_regulon:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Mar Regulon
        amp_lysis:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Amp Lysis
        initial_state_file:
          anyOf:
          - type: string
          - type: 'null'
          title: Initial State File
        skip_baseline:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Skip Baseline
        daughter_outdir:
          anyOf:
          - type: string
          - type: 'null'
          title: Daughter Outdir
        lineage_seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lineage Seed
        fail_at_max_duration:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Fail At Max Duration
        inherit_from:
          items:
            type: string
          type: array
          title: Inherit From
        spatial_environment_config:
          additionalProperties: true
          type: object
          title: Spatial Environment Config
        swap_processes:
          additionalProperties: true
          type: object
          title: Swap Processes
        flow:
          additionalProperties: true
          type: object
          title: Flow
        initial_state_overrides:
          items:
            type: string
          type: array
          title: Initial State Overrides
        initial_state:
          additionalProperties: true
          type: object
          title: Initial State
      type: object
      title: SimulationConfiguration
...
    UploadedAnalysisConfig:
      properties:
        config_id:
          type: string
          title: Config Id
      type: object
      required:
      - config_id
      title: UploadedAnalysisConfig
    UploadedSimulationConfig:
      properties:
        config_id:
          type: string
          title: Config Id
      type: object
      required:
      - config_id
      title: UploadedSimulationConfig
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError